feat!: Add ContentBlockLayoutConfiguration
To better accommodate the concept of page layouts, a new class has been introduced.
The class ContentBlockLayoutConfiguration contains data about the position and size of a ContentBlock.
These properties have been moved from ContentBlock to the new class. Instead, ContentBlock
now contains a reference to three ContentBlockLayoutConfiguration objects. These references
are used to describe the position and size of the ContentBlock in three different layouts:
Desktop, Tablet, and Mobile. This change is a breaking change. All tests that depended on the old
implementation of ContentBlock have been updated to reflect the new implementation.

BREAKING CHANGE: The properties of ContentBlock are completely updated