feat: Start adding API endpoints
This commit contains a lot of changes to the Application layer.
It was discovered that the previous design was flawed and didn't work well with
serialization from JSON to domain objects. This also caused issues with EF Core.
DTOs have been split into DTOs used for queries and DTOs used for commands.
A few test endpoints have been added. In the future, the rest of the application layer
should be updated and integrated with the API.
All validators have been cleared. New validators should be written based on the EF Core
configurations and domain rules.